Refactoring
The add_lyrics() method in Database class is not necessarily anymore.

The expired() method is model responsibility!

The testIfExist() method renamed to exists().
