+++ tests/test_database.py
@@ -16,31 +16,9 @@ class TestDataBase(TestBaseDb):
     def test_if_exist_lyrics(self):
         self._create()
 
-        answer = self.db.testIfExist(artist='Testudo', music='Test')
+        answer = self.db.exists(artist='Testudo', music='Test')
 
         self.assertTrue(answer)
-
-    def test_add_lyrics(self):
-        lyrics = self._create(commit=False)
-        self.db.addLyrics(lyrics)
-
-        result = Lyrics.select()
-        self.assertEqual(len(result), 1)
-
-    def test_expired(self):
-        new_date = datetime.strptime('2014-04-04', '%Y-%m-%d').date()
-        lyrics = self._create(created_date=new_date)
-
-        answer = self.db.testIfExpired(lyrics)
-
-        self.assertTrue(answer)
-
-    def test_not_expired(self):
-        lyrics = self._create()
-
-        answer = self.db.testIfExpired(lyrics)
-
-        self.assertFalse(answer)
 
     def test_update(self):
         lyrics = self._create()
